requirements.txt
discord.py==2.2.2
elevenlabs==0.2.7
pydub==0.25.1
python-dotenv==1.0.0
Requests==2.29.0
PyNaCl